requirements.txt
httplib2
Pillow
lxml
bs4
